Movement fixed

+++ PratformUnit.h
@@ -8,7 +8,7 @@ public:
 	int currentSpriteIndex;
 	int counter;
 
-	PratformUnit(std::vector<Sprite*> sprites, int x, int y, int width, int height);
+	PratformUnit(std::vector<Sprite*> sprites, double x, double y, double width, double height);
 	virtual void update();
 	void increase(double ñoef);
 	void decrease(double ñoef);
